Moved command line interface code to proper location
