+++ ImaginationServer.SingleServer/packages.config
@@ -4,6 +4,7 @@
   <package id="EntityFramework" version="6.0.0" targetFramework="net461" />
   <package id="FluentNHibernate" version="2.0.3.0" targetFramework="net461" />
   <package id="Iesi.Collections" version="4.0.0.4000" targetFramework="net46" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
   <package id="NHibernate" version="5.0.0" targetFramework="net461" />
   <package id="Remotion.Linq" version="2.1.2" targetFramework="net461" />
   <package id="Remotion.Linq.EagerFetching" version="2.1.0" targetFramework="net461" />
